# generated-swagger-docs.json
{"swagger":"2.0","info":{"title":"Personal Book Commenter API","description":"Testing the Fastify API with swagger.","version":"1.0.0"},"definitions":{"def-0":{"$schema":"http://json-schema.org/schema","type":"object","required":["email","password","fullName"],"additionalProperties":false,"properties":{"fullName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"loginToken":{"type":"string"}},"title":"users"},"def-1":{"$schema":"http://json-schema.org/schema","type":"object","required":["email","password"],"additionalProperties":false,"properties":{"email":{"type":"string"},"password":{"type":"string"}},"title":"login"},"def-2":{"$schema":"http://json-schema.org/schema","type":"object","required":["title","author","realeaseYear","genre","pages"],"additionalProperties":false,"properties":{"title":{"type":"string"},"author":{"type":"string"},"realeaseYear":{"type":"number"},"genre":{"type":"string","enum":["FICTION","NON_FICTION","DRAMA","THRILLER","SCI_FI","OTHERS"]},"pages":{"type":"number"},"isbn":{"type":"string"},"rating":{"type":"string","enum":["ONE","TWO","THREE","FOUR","FIVE"]},"user":{"type":"object"}},"title":"bookings"},"def-3":{"$schema":"http://json-schema.org/schema","type":"object","required":["chapter"],"additionalProperties":false,"properties":{"chapter":{"type":"number"},"description":{"type":"string"},"book":{"type":"string"}},"title":"chapetrs"},"def-4":{"$schema":"http://json-schema.org/schema","type":"object","required":["comment"],"additionalProperties":false,"properties":{"comment":{"type":"string"}},"title":"comment"},"def-5":{"$schema":"http://json-schema.org/schema","type":"object","additionalProperties":false,"required":["id"],"properties":{"id":{"type":"integer"}},"title":"paraming"},"def-6":{"$schema":"http://json-schema.org/schema","type":"object","additionalProperties":false,"required":["bookId","id"],"properties":{"bookId":{"type":"integer"},"id":{"type":"integer"}},"title":"paramingTwice"},"def-7":{"$schema":"http://json-schema.org/schema","type":"object","additionalProperties":false,"required":["bookId","id","commentId"],"properties":{"bookId":{"type":"integer"},"id":{"type":"integer"},"commentId":{"type":"integer"}},"title":"ThreeIds"},"def-8":{"$schema":"http://json-schema.org/schema","type":"object","required":["characterName"],"additionalProperties":false,"properties":{"characterName":{"type":"string"},"description":{"type":"string"},"book":{"type":"string"}},"title":"personas"}},"paths":{"/register":{"post":{"tags":["auth","public"],"parameters":[{"name":"body","in":"body","schema":{"$schema":"http://json-schema.org/schema","$id":"todo.com","type":"object","required":["email","password","fullName"],"additionalProperties":false,"properties":{"fullName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"loginToken":{"type":"string"}}}}],"responses":{"200":{"description":"Default Response"}}}},"/login":{"post":{"tags":["auth","public"],"parameters":[{"name":"body","in":"body","schema":{"$schema":"http://json-schema.org/schema","$id":"todo.com","type":"object","required":["email","password"],"additionalProperties":false,"properties":{"email":{"type":"string"},"password":{"type":"string"}}}}],"responses":{"200":{"description":"Default Response"}}}},"/logout":{"get":{"tags":["auth","private"],"responses":{"200":{"description":"Default Response"}}}},"/books":{"get":{"tags":["public","book"],"responses":{"200":{"description":"Default Response"}}}},"/books/{id}":{"get":{"tags":["public","book"],"parameters":[{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}}},"/me":{"get":{"tags":["private"],"responses":{"200":{"description":"Default Response"}}}},"/me/books":{"get":{"tags":["private","book"],"responses":{"200":{"description":"Default Response"}}},"post":{"tags":["book","private"],"parameters":[{"name":"body","in":"body","schema":{"$schema":"http://json-schema.org/schema","$id":"todo.com","type":"object","required":["title","author","realeaseYear","genre","pages"],"additionalProperties":false,"properties":{"title":{"type":"string"},"author":{"type":"string"},"realeaseYear":{"type":"number"},"genre":{"type":"string","enum":["FICTION","NON_FICTION","DRAMA","THRILLER","SCI_FI","OTHERS"]},"pages":{"type":"number"},"isbn":{"type":"string"},"rating":{"type":"string","enum":["ONE","TWO","THREE","FOUR","FIVE"]},"user":{"type":"object"}}}}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{id}":{"get":{"tags":["private","book"],"parameters":[{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}},"put":{"tags":["book","private"],"parameters":[{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}},"delete":{"tags":["book","private"],"parameters":[{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{id}/chapters":{"post":{"tags":["private"],"parameters":[{"name":"body","in":"body","schema":{"$schema":"http://json-schema.org/schema","$id":"todo.com","type":"object","required":["chapter"],"additionalProperties":false,"properties":{"chapter":{"type":"number"},"description":{"type":"string"},"book":{"type":"string"}}}},{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{bookId}/chapters/{id}":{"put":{"tags":["private"],"parameters":[{"type":"integer","required":true,"in":"path","name":"bookId"},{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}},"delete":{"tags":["private"],"parameters":[{"type":"integer","required":true,"in":"path","name":"bookId"},{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{bookId}/chapters/{id}/comment":{"post":{"tags":["private"],"parameters":[{"name":"body","in":"body","schema":{"$schema":"http://json-schema.org/schema","$id":"todo.com","type":"object","required":["comment"],"additionalProperties":false,"properties":{"comment":{"type":"string"}}}},{"type":"integer","required":true,"in":"path","name":"bookId"},{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{bookId}/chapters/{id}/comment/{commentId}":{"delete":{"tags":["private"],"parameters":[{"type":"integer","required":true,"in":"path","name":"bookId"},{"type":"integer","required":true,"in":"path","name":"id"},{"type":"integer","required":true,"in":"path","name":"commentId"}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{id}/personas":{"post":{"tags":["private"],"parameters":[{"name":"body","in":"body","schema":{"$schema":"http://json-schema.org/schema","$id":"todo.com","type":"object","required":["characterName"],"additionalProperties":false,"properties":{"characterName":{"type":"string"},"description":{"type":"string"},"book":{"type":"string"}}}},{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{bookId}/personas/{id}":{"put":{"tags":["private"],"parameters":[{"type":"integer","required":true,"in":"path","name":"bookId"},{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}},"delete":{"tags":["private"],"parameters":[{"type":"integer","required":true,"in":"path","name":"bookId"},{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{bookId}/personas/{id}/comment":{"post":{"tags":["private"],"parameters":[{"name":"body","in":"body","schema":{"$schema":"http://json-schema.org/schema","$id":"todo.com","type":"object","required":["comment"],"additionalProperties":false,"properties":{"comment":{"type":"string"}}}},{"type":"integer","required":true,"in":"path","name":"bookId"},{"type":"integer","required":true,"in":"path","name":"id"}],"responses":{"200":{"description":"Default Response"}}}},"/me/books/{bookId}/personas/{id}/comment/{commentId}":{"delete":{"tags":["private"],"parameters":[{"type":"integer","required":true,"in":"path","name":"bookId"},{"type":"integer","required":true,"in":"path","name":"id"},{"type":"integer","required":true,"in":"path","name":"commentId"}],"responses":{"200":{"description":"Default Response"}}}}},"host":"localhost:3000","schemes":["http"],"consumes":["application/json"],"produces":["application/json"],"tags":[{"name":"public","description":"Public routes that do not require authentication"},{"name":"auth","description":"Routes for Authentication"},{"name":"private","description":"Authentication routes for private user operations"},{"name":"book","description":"Book  routes, both public and private"}]}